feat: Introduce FormVocabularyService and refactor related components
- Added FormVocabularyService to handle form lookup vocabularies and typeahead sources.
- Updated FormVocabularyController to utilize FormVocabularyService instead of VocabService.
- Refactored UsersService tests to accommodate the new FormVocabularyService.
- Created comprehensive tests for FormVocabularyService covering various functionalities.
- Removed deprecated VocabService references and tests.
- Updated documentation to reflect changes in service architecture and usage.
