+++ packages/redbox-core/src/bootstrap.ts
@@ -80,8 +80,6 @@ export async function coreBootstrap(): Promise<void> {
     }
 
     sails.log.verbose("Forms service, bootstrapped.");
-    await lastValueFrom(sails.services.vocabservice.bootstrap() as Observable<unknown>);
-    sails.log.verbose("Vocab service, bootstrapped.");
     await sails.services.vocabularyservice.bootstrapData();
     sails.log.verbose("Vocabulary bootstrap data, loaded.");
     await sails.services.recordsservice.bootstrapData();
