+++ packages/redbox-core/src/services/index.ts
@@ -16,6 +16,7 @@ import * as DashboardTypesServiceModule from './DashboardTypesService';
 import * as DoiServiceModule from './DoiService';
 import * as EmailServiceModule from './EmailService';
 import * as FigshareServiceModule from './FigshareService';
+import * as FormVocabularyServiceModule from './FormVocabularyService';
 import * as FormRecordConsistencyServiceModule from './FormRecordConsistencyService';
 import * as FormsServiceModule from './FormsService';
 import * as IntegrationAuditServiceModule from './IntegrationAuditService';
@@ -39,7 +40,6 @@ import * as TranslationServiceModule from './TranslationService';
 import * as TriggerServiceModule from './TriggerService';
 import * as UsersServiceModule from './UsersService';
 import * as ViewUtilsServiceModule from './ViewUtilsService';
-import * as VocabServiceModule from './VocabService';
 import * as VocabularyServiceModule from './VocabularyService';
 import * as WorkflowStepsServiceModule from './WorkflowStepsService';
 import * as WorkspaceAsyncServiceModule from './WorkspaceAsyncService';
@@ -62,6 +62,7 @@ export { DashboardTypesServiceModule as DashboardTypesService };
 export { DoiServiceModule as DoiService };
 export { EmailServiceModule as EmailService };
 export { FigshareServiceModule as FigshareService };
+export { FormVocabularyServiceModule as FormVocabularyService };
 export { FormRecordConsistencyServiceModule as FormRecordConsistencyService };
 export { FormsServiceModule as FormsService };
 export { IntegrationAuditServiceModule as IntegrationAuditService };
@@ -85,7 +86,6 @@ export { TranslationServiceModule as TranslationService };
 export { TriggerServiceModule as TriggerService };
 export { UsersServiceModule as UsersService };
 export { ViewUtilsServiceModule as ViewUtilsService };
-export { VocabServiceModule as VocabService };
 export { VocabularyServiceModule as VocabularyService };
 export { WorkflowStepsServiceModule as WorkflowStepsService };
 export { WorkspaceAsyncServiceModule as WorkspaceAsyncService };
@@ -156,6 +156,11 @@ export const ServiceExports = {
   },
   get FigshareService() {
     return getOrCreateService('FigshareService', () => new FigshareServiceModule.Services.FigshareService().exports());
+  },
+  get FormVocabularyService() {
+    return getOrCreateService('FormVocabularyService', () =>
+      new FormVocabularyServiceModule.Services.FormVocabulary().exports()
+    );
   },
   get IntegrationAuditService() {
     return getOrCreateService('IntegrationAuditService', () => new IntegrationAuditServiceModule.Services.IntegrationAuditService().exports());
@@ -241,9 +246,6 @@ export const ServiceExports = {
   },
   get ViewUtilsService() {
     return getOrCreateService('ViewUtilsService', () => new ViewUtilsServiceModule.Services.ViewUtils().exports());
-  },
-  get VocabService() {
-    return getOrCreateService('VocabService', () => new VocabServiceModule.Services.Vocab().exports());
   },
   get VocabularyService() {
     return getOrCreateService('VocabularyService', () =>
